requirements.txt
# ───────────────────────────────────────────
# 유튜브 트렌드 주간 리포트 자동화
# 설치: pip install -r requirements.txt
# ───────────────────────────────────────────

# 환경 변수
python-dotenv>=1.0.1

# YouTube Data API v3 + Gmail API
google-api-python-client>=2.100.0
google-auth-oauthlib>=1.1.0
google-auth-httplib2>=0.2.0

# Claude API
anthropic>=0.37.0

# Notion
notion-client>=2.2.1

# PDF 생성
fpdf2>=2.7.9

# 차트 이미지
matplotlib>=3.8.0
